push

+++ core/src/main/java/de/frinshhd/core/Ban.java
@@ -6,7 +6,7 @@ import java.math.BigInteger;
 import java.util.UUID;
 
 public class Ban {
-    public BigInteger id;
+    public int id;
 
     public UUID uuid;
